requirements.txt
Jinja2
MarkupSafe
PyPika
PyYAML
Pygments
SQLAlchemy
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
async-timeout
attrs
backoff
bcrypt
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
colorama
coloredlogs
cryptography
dataclasses-json
distro
durationpy
exceptiongroup
fastapi
filelock
filetype
flatbuffers
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-auth
googleapis-common-protos
greenlet
grpcio-status
grpcio
h11
httpcore
httptools
httpx-sse
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
joblib
jsonpatch
jsonpointer
jsonschema-specifications
jsonschema
kubernetes
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-text-splitters
langchain
langchain_community
langchain_huggingface
langgraph
langsmith
markdown-it-py
marshmallow
mdurl
mmh3
mpmath
multidict
mypy_extensions
networkx
numpy
oauthlib
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pandas
pillow
posthog
propcache
proto-plus
protobuf
pulsar-client
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic-settings
pydantic
pydantic_core
pylance
pypdf
pyproject_hooks
pyreadline3
python-dateutil
python-dotenv
pytz
referencing
regex
requests-oauthlib
requests-toolbelt
requests
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
shellingham
six
sniffio
starlette
sympy
tenacity
threadpoolctl
tokenizers
tomli
torch
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
watchfiles
websocket-client
websockets
wrapt
yarl
zipp
zstandard
fastapi
uvicorn[standard]
websockets
streamlit
python-multipart
